Changed some settings for UI controller input and changed the chapter selection navigation.
Changed some settings for UI controller input, now it works better to use the joystick on the controller. The chapter navigation is now no longer looping through all chapter buttons, instead it is looping on the left and right columns, maybe that's better? The old code is still in the script if I want to go back.