All one shot sounds are added in FMOD and works.
The FMOD events for sounds that Rémi made are in a folder called temporary, to make it easier to find them to replace them.
Volumes are set to something that works.
Sound manager script is not fully cleaned up and no parameters are added to modify volumes, pitch and so based on different things.
No old Unity sound stuff is removed yet. Like audio sources.