Fixed the bug where the Thintel font wouldn't show up while playing not in fullscreen on mac.
It is the second time I fixed this, I wasn't happy with the first solution because it changed how the font behaved both in the editor and when the game was built. Now the font material settings are changed back to what they used to be and the bug is fixed by changing the resolution for one frame.